Reorganized the file structure a bit
Separated mergesort and insertionsort into separate files
